Autorização por perfis

+++ src/Back-End/ProjetoGerenciar/Controllers/UserController.cs
@@ -19,14 +19,14 @@ public class UserController : ControllerBase
     {
         _context = context;
     }
-
     [HttpGet]
+    [AllowAnonymous]
     public async Task<IEnumerable<User>> Get()
     {
         return await _context.Users.Find(_ => true).ToListAsync();
     }
-
     [HttpGet("{id}")]
+    [AllowAnonymous]
     public async Task<ActionResult<User>> GetById(string id)
     {
         var user = await _context.Users.Find(p => p.Id == id).FirstOrDefaultAsync();
@@ -35,10 +35,8 @@ public class UserController : ControllerBase
         {
             return NotFound();
         }
-
         return user;
     }
-
     [AllowAnonymous]
     [HttpPost]
     public async Task<ActionResult<User>> Create(UserDto user)
@@ -49,8 +47,6 @@ public class UserController : ControllerBase
         {
             return BadRequest("E-mail já cadastrado no sistema");
         }
-
-
         User novoUser = new User()
         {
             Id = user.Id,
@@ -59,49 +55,65 @@ public class UserController : ControllerBase
             Email = user.Email,
             Perfil = user.Perfil
         };
-
-       
-
         await _context.Users.InsertOneAsync(novoUser);
         return CreatedAtRoute(new { id = novoUser.Id }, novoUser);
     }
-
     [HttpPut("{id}")]
+    [Authorize(Roles = "AdminRh")]
     public async Task<IActionResult> Update(string id, UserDto userIn)
     {
         var user = await _context.Users.Find(u => u.Id == id).FirstOrDefaultAsync();
-
-        if (user == null)
+        try
         {
-            return NotFound();
+            if (user == null)
+            {
+                return NotFound();
+            }
+            var userHasPermission = User.IsInRole("AdminRh");
+            if (!userHasPermission)
+            {
+                return Unauthorized("Usuário não autorizado");
+            }
+            User novoUser = new User()
+            {
+                Nome = user.Nome,
+                Senha = BCrypt.Net.BCrypt.HashPassword(user.Senha),
+                Email = user.Email,
+                Perfil = user.Perfil
+            };
+            await _context.Users.ReplaceOneAsync(u => u.Id == id, novoUser);
+            return NoContent();
         }
-        User novoUser = new User()
+        catch (UnauthorizedAccessException)
         {
-            Nome = user.Nome,
-            Senha = BCrypt.Net.BCrypt.HashPassword(user.Senha),
-            Email = user.Email,
-            Perfil = user.Perfil
-        };
-
-        await _context.Users.ReplaceOneAsync(u => u.Id == id, novoUser);
-
-        return NoContent();
+            return Unauthorized("Usuário não autorizado");
+        }
     }
-
-    [AllowAnonymous]
     [HttpDelete("{id}")]
+    [Authorize(Roles = "AdminRh")]
     public async Task<IActionResult> Delete(string id)
     {
-        var user = await _context.Users.Find(u => u.Id == id).FirstOrDefaultAsync();
-
-        if (user == null)
+        try
         {
-            return NotFound();
+            var user = await _context.Users.Find(u => u.Id == id).FirstOrDefaultAsync();
+
+            if (user == null)
+            {
+                return NotFound();
+            }
+            var userHasPermission = User.IsInRole("AdminRh");
+
+            if (!userHasPermission)
+            {
+                return Unauthorized("Usuário não autorizado");
+            }
+            await _context.Users.DeleteOneAsync(u => u.Id == id);
+            return NoContent();
         }
-
-        await _context.Users.DeleteOneAsync(u => u.Id == id);
-
-        return NoContent();
+        catch (UnauthorizedAccessException)
+        {
+            return Unauthorized("Usuário não autorizado");
+        }
     }
 
     [AllowAnonymous]
@@ -109,19 +121,13 @@ public class UserController : ControllerBase
     public async Task<IActionResult> Authenticate(AuthenticateDto model)
     {
         var user = await _context.Users.Find(e => e.Email == model.Email).FirstOrDefaultAsync();
-
         if (user == null || !BCrypt.Net.BCrypt.Verify(model.Senha, user.Senha))
         {
             return Unauthorized();
         }
-
         var jwt = GenerateJwtToken(user);
-
-
-
         return Ok(new { jwtToken = jwt });
     }
-
     private string GenerateJwtToken(User model)
     {
         var tokenHandler = new JwtSecurityTokenHandler();
@@ -140,6 +146,4 @@ public class UserController : ControllerBase
         var token = tokenHandler.CreateToken(TokenDescriptor);
         return tokenHandler.WriteToken(token);
     }
-
-
 }
